add oracle post apis to swagger
