+++ x/shield/module.go
@@ -55,7 +55,7 @@ type AppModule struct {
 }
 
 // NewAppModule creates a new AppModule object.
-func NewAppModule(keeper Keeper, accountKeeper types.AccountKeeper, 
+func NewAppModule(keeper Keeper, accountKeeper types.AccountKeeper,
 	stakingKeeper types.StakingKeeper, supplyKeeper types.SupplyKeeper) AppModule {
 	return AppModule{
 		AppModuleBasic: AppModuleBasic{},
